【AC】[AtCoder Beginner Contest 059]D - Alice&Brown

+++ abc059/arc072_b.cpp
@@ -53,7 +53,7 @@ bool solve(int y, int x, bool turn) {
 }
 
 int main() {
-    int a, b;
+    ll a, b;
     cin >> a >> b;
 //    bool ok = solve(a, b, true);
     bool ok = abs(a - b) > 1;
